# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PR/PO Approval System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /approvals/{id}/approve:
    post:
      operationId: approvePR
      tags: [approvals]
      summary: Approve a PR
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalActionRequest"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
  /approvals/{id}/reject:
    post:
      operationId: rejectPR
      tags: [approvals]
      summary: Reject a PR
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalActionRequest"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
  /approval-rules:
    get:
      operationId: getApprovalRules
      tags: [approvals]
      summary: Get approval rules
      responses:
        "200":
          description: Approval rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApprovalRule"

    post:
      operationId: createApprovalRule
      tags: [approvals]
      summary: Create approval rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApprovalRuleRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRule"
  /approval-rules/{id}:
    put:
      operationId: updateApprovalRule
      tags: [approvals]
      summary: Update approval rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApprovalRuleRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRule"

    delete:
      operationId: deleteApprovalRule
      tags: [approvals]
      summary: Delete approval rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get notifications for current user
      parameters:
        - name: unread
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
  /notifications/{id}/read:
    put:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        email:
          type: string
        department:
          type: string
        position:
          type: string
        role:
          type: string
          enum: [admin, user, approver, purchasing]
        superiorId:
          type: integer
          nullable: true
        superiorName:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - name
        - department
        - position
        - role
        - isActive
        - createdAt
...
    Approval:
      type: object
      properties:
        id:
          type: integer
        prId:
          type: integer
        approverId:
          type: integer
        approverName:
          type: string
        level:
          type: integer
        status:
          type: string
          enum: [pending, approved, rejected]
        notes:
          type: string
          nullable: true
        actionAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - prId
        - approverId
        - approverName
        - level
        - status
        - createdAt
...
    ApprovalActionRequest:
      type: object
      properties:
        notes:
          type: string
          nullable: true
      required: []
    ApprovalRule:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        minAmount:
          type: number
        maxAmount:
          type: number
          nullable: true
        levels:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalRuleLevel"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - minAmount
        - levels
        - createdAt
    ApprovalRuleLevel:
      type: object
      properties:
        id:
          type: integer
        ruleId:
          type: integer
        level:
          type: integer
        role:
          type: string
          enum: [admin, user, approver, purchasing]
        position:
          type: string
          nullable: true
      required:
        - id
        - ruleId
        - level
        - role
    CreateApprovalRuleRequest:
      type: object
      properties:
        name:
          type: string
        minAmount:
          type: number
        maxAmount:
          type: number
          nullable: true
        levels:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalRuleLevelInput"
      required:
        - name
        - minAmount
        - levels
    ApprovalRuleLevelInput:
      type: object
      properties:
        level:
          type: integer
        role:
          type: string
          enum: [admin, user, approver, purchasing]
        position:
          type: string
          nullable: true
      required:
        - level
        - role
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [approval_request, approved, rejected, received, info]
        isRead:
          type: boolean
        prId:
          type: integer
          nullable: true
        poId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - message
        - type
        - isRead
        - createdAt
    NotificationListResponse:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        unreadCount:
          type: integer
        total:
          type: integer
      required:
        - notifications
        - unreadCount
        - total
...