# lib/api-zod/src/generated/types/approval.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PR/PO Approval System API
 * OpenAPI spec version: 0.1.0
 */
import type { ApprovalStatus } from "./approvalStatus";

export interface Approval {
  id: number;
  prId: number;
  approverId: number;
  approverName: string;
  level: number;
  status: ApprovalStatus;
  notes?: string | null;
  actionAt?: Date | null;
  createdAt: Date;
}